pyproject.toml
[tool.poetry]
name = "aftershoot-wb-prediction"
version = "1.0.0"
description = "Professional ML solution for Aftershoot White Balance prediction from TIFF images"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
torch = "^2.1.0"
torchvision = "^0.16.0"
torchaudio = "^2.1.0"
timm = "0.9.12"
albumentations = "1.3.1"
opencv-python = "4.8.1.78"
pandas = "2.1.4"
numpy = "1.24.4"
scikit-learn = "1.3.2"
matplotlib = "3.8.2"
seaborn = "0.13.0"
tqdm = "4.66.1"
Pillow = "10.1.0"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
ipykernel = "^6.25.0"
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[tool.poetry.group.colab.dependencies]
# Colab-specific packages
google-colab = {version = "*", optional = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
train = "main:main"
eda = "main:main"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]